apps/backend/pyproject.toml
[tool.pdm.dev-dependencies]
dev = []

[tool.pdm.build]
includes = []

[tool.uv]
dev-dependencies = [
    "pytest-cov>=6.2.1",
    "radon>=6.0.1",
    "ty>=0.0.1a17",
    "wily>=1.12.2",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
authors = [
    { name = "Robert Meissner", email = "robert@robertforpresent.de" },
]
requires-python = "<4.0,>=3.10"
dependencies = [
    "pyperclip<2.0.0,>=1.9.0",
    "odfpy<2.0.0,>=1.4.1",
    "python-dotenv<1.0.0,>=0.21.0",
    "httpx<1.0.0,>=0.27.2",
    "matplotlib<4.0.0,>=3.6.2",
    "pre-commit<3.0.0,>=2.20.0",
    "streamlit<2.0.0,>=1.37.1",
    "tqdm<5.0.0,>=4.66.5",
    "pytest<9.0.0,>=8.3.2",
    "fastapi<1.0.0,>=0.114.0",
    "uvicorn<1.0.0,>=0.30.6",
    "pandas>=2.3.1",
    "numpy==1.26.3",
    "bs4<1.0.0,>=0.0.2",
    "pyarrow<18.0.0,>=17.0.0",
    "detect-secrets>=1.5.0",
]
name = "Game_Overview"
version = "0.1.0"
description = ""
